feat: Add GUI functionality for running tasks
The code changes introduce GUI functionality for running tasks. This includes creating a control panel window with a "Run" button and implementing the corresponding callback function. The commit message follows the established convention of using a prefix to indicate the type of change, in this case, "feat" for a new feature.