feat: Update main.cpp to support command line arguments for initial and final file paths
The code changes modify the main.cpp file to support command line arguments for specifying the initial and final file paths. This allows the user to provide the file paths when running the program, instead of hardcoding them in the code. The commit message follows the established convention of using a prefix to indicate the type of change, in this case, "feat" for a new feature.