feat: Enable exporting to OBJ format
The code changes add the ability to export the 3D model to the OBJ format. A new button, "Export to OBJ", is added to the HTML file, and a corresponding JavaScript function, `exportToObj()`, is implemented in the `sketch.js` file. The `exportToObj()` function generates the OBJ data for the 3D model and provides a download link for the OBJ file. This feature allows users to export the model and use it in other 3D modeling software.